Trying to complete the user management for authorization sections, running into an issue with the DB context Create. Run project for error message.
ArgumentException: AddDbContext was called with configuration, but the context type 'project_trackerContext' only declares a parameterless constructor. This means that the configuration passed to AddDbContext will never be used. If configuration is passed to AddDbContext, then 'project_trackerContext' should declare a constructor that accepts a DbContextOptions<project_trackerContext> and must pass it to the base constructor for DbContext.
Microsoft.Extensions.DependencyInjection.EntityFrameworkServiceCollectionExtensions.CheckContextConstructors<TContext>()